refactor: step2 리뷰 반영
- Composable 외부로 공개되지 않는 경우 접근제어자 private으로 변경
- Test코드 SignUpTextField랑 동일한 패키지로 이동
